refactor(cruzeiros): split landing sections
